[frontend] extract fragment entitySettings from the component
